Add initial functionality of the calculator
The calculator can currenrtly perform the add, substract, multiply, and divide functions correctly (two numbers are evaluated to the right answer according to the operator).
There remains two bugs: 1) The calculator still needs to not evaluate more than a single pair of numnbers at a time 2) When a result is displayed, pressing a new digit should clear the result and start a new calculation instead of apppending the digit to the resut.